First commit of working gen3-fit-fftw-2-init.
This commit is the first of gen3-fit-fftw-2-init.  It successfully compiles, and is used to perform all the computations and actions that must be performed only once per experiment in FFTW fitting of the FITS files.

A modified Makefile was created that uses the headers from this project and renames the output to be .o.

 On branch master
 Changes to be committed:

	modified:   Makefile
	new file:   gen3-fftw-ffit-2-init.o
	modified:   gen3-fftw-fit-2-init.cpp
	new file:   gen3-fftw-fit-2.o
	modified:   gen3pp/fftwfit.h